load analytics eagerly
