+++ scripts/scripts.js
@@ -12,6 +12,7 @@ import {
   loadCSS,
   fetchPlaceholders,
   getMetadata,
+  loadScript,
 } from './aem.js';
 import { picture, source, img } from './dom-helpers.js';
 
@@ -22,6 +23,8 @@ import {
   setPageLanguage,
   cookiePopUp,
   showCookieConsent,
+  isInternalPage,
+  scriptEnabled,
   PATH_PREFIX,
 } from './utils.js';
 
@@ -236,12 +239,25 @@ export async function load404() {
   return null;
 }
 
+async function loadAdobeLaunch() {
+  if (!scriptEnabled()) { return; }
+
+  const config = await fetchPlaceholders(PATH_PREFIX);
+  const env = config.environment || 'Dev';
+  await loadScript(config[`analyticsEndpoint${env}`]);
+}
+
 async function loadEager(doc) {
   setPageLanguage();
   decorateTemplateAndTheme();
   await createSkipToMainNavigationBtn();
   await cookiePopUp();
   renderWBDataLayer();
+
+  if (!isInternalPage()) {
+    await loadAdobeLaunch();
+  }
+
   const main = doc.querySelector('main');
   if (main) {
     decorateMain(main);
